requirements.txt
# Interactive Brokers
ib_insync==0.9.86

# Crypto
ccxt==4.2.0

# Dados históricos
yfinance==0.2.36
pandas==2.1.4
numpy==1.26.3

# Banco de dados
psycopg2-binary==2.9.9
sqlalchemy==2.0.25

# Backtesting
vectorbt==0.26.1

# Análise
scipy==1.11.4
statsmodels==0.14.1
scikit-learn==1.3.2

# Visualização
matplotlib==3.8.2
plotly==5.18.0

# Utilitários
python-dotenv==1.0.0
pyyaml==6.0.1
loguru==0.7.2
schedule==1.2.1

# Jupyter
jupyter==1.0.0
ipykernel==6.28.0
